[188089] add `LogWriter` module `use`d to write logs
