+++ core/sup_tree_gear/gear_log_writer.ex
@@ -12,24 +12,17 @@ defmodule AntikytheraCore.GearLog.Writer do
   After each successful log rotation, old log file is uploaded to cloud storage.
   """
 
-  use GenServer
+  use AntikytheraCore.LogWriter, [
+    rotate_interval:         (if Mix.env() == :test, do: 500, else: 7_200_000),
+    additional_state_fields: [min_level: Level, uploader: Croma.TypeGen.nilable(Croma.Pid)],
+  ]
   alias Antikythera.{Time, ContextId, GearName}
-  alias AntikytheraCore.GearLog.{FileHandle, Level, ContextHelper}
+  alias AntikytheraCore.GearLog.{Level, ContextHelper}
   alias AntikytheraCore.Config.Gear, as: GearConfig
   alias AntikytheraCore.Ets.ConfigCache
   alias AntikytheraCore.Alert.Manager, as: CoreAlertManager
   require AntikytheraCore.Logger, as: L
   alias AntikytheraEal.LogStorage
-
-  defmodule State do
-    use Croma.Struct, recursive_new?: true, fields: [
-      min_level:   Level,
-      file_handle: Croma.Tuple, # FileHandle.t
-      empty?:      Croma.Boolean,
-      timer:       Croma.Reference,
-      uploader:    Croma.TypeGen.nilable(Croma.Pid),
-    ]
-  end
 
   def start_link([gear_name, logger_name]) do
     opts = if logger_name, do: [name: logger_name], else: []
@@ -48,17 +41,10 @@ defmodule AntikytheraCore.GearLog.Writer do
   end
 
   @impl true
-  def handle_cast({_, level, _, _} = gear_log,
-                  %State{min_level: min_level, file_handle: handle, timer: timer} = state) do
+  def handle_cast({_, level, _, _} = gear_log, %State{min_level: min_level} = state) do
     if Level.write_to_log?(min_level, level) do
-      case FileHandle.write(handle, gear_log) do
-        {:kept_open, new_handle} -> {:noreply, %State{state | file_handle: new_handle, empty?: false}}
-        {:rotated  , new_handle} ->
-          # Log file is just rotated as its size has exceeded the upper limit.
-          # Note that the current message is written to the newly-opened log file and thus it's not empty.
-          new_timer = arrange_next_rotation(timer)
-          {:noreply, %State{state | file_handle: new_handle, empty?: false, timer: new_timer}}
-      end
+      new_state = write_log(state, gear_log)
+      {:noreply, new_state}
     else
       {:noreply, state}
     end
@@ -80,35 +66,8 @@ defmodule AntikytheraCore.GearLog.Writer do
   end
 
   @impl true
-  def handle_info(:rotate, state) do
-    {:noreply, rotate(state)}
-  end
   def handle_info({:DOWN, _ref, :process, _pid, _reason}, state) do
     {:noreply, %State{state | uploader: nil}}
-  end
-
-  defp rotate(%State{file_handle: handle, empty?: empty?, timer: timer} = state0) do
-    new_timer = arrange_next_rotation(timer)
-    state1 = %State{state0 | timer: new_timer}
-    if empty? do
-      state1
-    else
-      %State{state1 | file_handle: FileHandle.rotate(handle), empty?: true}
-    end
-  end
-
-  @impl true
-  def terminate(_reason, %State{file_handle: handle}) do
-    FileHandle.close(handle)
-  end
-
-  @rotate_interval (if Mix.env() == :test, do: 500, else: 7_200_000)
-
-  defp arrange_next_rotation(timer) do
-    if timer do
-      Process.cancel_timer(timer)
-    end
-    Process.send_after(self(), :rotate, @rotate_interval)
   end
 
   #
